requirements.txt
# Flask and Web
Flask==3.0.0
Werkzeug==3.0.1

# AI and ML - Core
tensorflow>=2.15.0
keras>=2.15.0
numpy>=1.24.0
opencv-python>=4.9.0
Pillow>=10.0.0
matplotlib>=3.8.0

# LangChain
langchain>=0.1.0
langchain-groq>=0.3.0
langchain-community>=0.0.12

# Neo4j
neo4j>=5.16.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.2
requests>=2.31.0
